Remove "impossible" condition
Under normal circumstances it should never trigger, why?

In fact we can crash here if the condition passes somehow.

First check if the instance is in "no fun mode".
If it fails then skip the rest of the condition and we get no crash.

If it passes then it will check for an empty "api key".

If it fails then skip the "has password" check and we get no crash.
If it passes then check for "has password" but that means reading the global "server info" which is null so we crash.

This last conditions should never happen because there is no way to get the "server info" in "no fun mode" without the key in the first place.